feat: new console options and code refactoring
