+++ src/Exceptions/CodeGenerateCommandException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace DevLnk\LaravelCodeBuilder\Exceptions;
+
+use Exception;
+
+class CodeGenerateCommandException extends Exception
+{
+}
