+++ src/Services/Builders/Advanced/TSBuilder.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DevLnk\LaravelCodeBuilder\Services\Builders\Advanced;
+
+use DevLnk\LaravelCodeBuilder\Services\Builders\AbstractBuilder;
+
+class TSBuilder extends AbstractBuilder
+{
+    public function build(): void
+    {
+        // TODO ts builder
+    }
+}
