+++ src/Services/Builders/BuildFactory.php
@@ -7,6 +7,14 @@ namespace DevLnk\LaravelCodeBuilder\Services\Builders;
 use DevLnk\LaravelCodeBuilder\Enums\BuildType;
 use DevLnk\LaravelCodeBuilder\Exceptions\NotFoundBuilderException;
 use DevLnk\LaravelCodeBuilder\Exceptions\NotFoundCodePathException;
+use DevLnk\LaravelCodeBuilder\Services\Builders\Core\AddActionBuilder;
+use DevLnk\LaravelCodeBuilder\Services\Builders\Core\ControllerBuilder;
+use DevLnk\LaravelCodeBuilder\Services\Builders\Core\DTOBuilder;
+use DevLnk\LaravelCodeBuilder\Services\Builders\Core\EditActionBuilder;
+use DevLnk\LaravelCodeBuilder\Services\Builders\Core\FormBuilder;
+use DevLnk\LaravelCodeBuilder\Services\Builders\Core\ModelBuilder;
+use DevLnk\LaravelCodeBuilder\Services\Builders\Core\RequestBuilder;
+use DevLnk\LaravelCodeBuilder\Services\Builders\Core\RouteBuilder;
 use DevLnk\LaravelCodeBuilder\Services\CodePath\CodePath;
 use DevLnk\LaravelCodeBuilder\Services\CodeStructure\CodeStructure;
 use Illuminate\Contracts\Filesystem\FileNotFoundException;
