+++ src/Services/CodePath/Core/ControllerPath.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DevLnk\LaravelCodeBuilder\Services\CodePath\Core;
+
+use DevLnk\LaravelCodeBuilder\Services\CodePath\AbstractPath;
+
+readonly class ControllerPath extends AbstractPath
+{
+
+}
