+++ src/Services/CodePath/Core/EditActionPath.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DevLnk\LaravelCodeBuilder\Services\CodePath\Core;
+
+use DevLnk\LaravelCodeBuilder\Services\CodePath\AbstractPath;
+
+readonly class EditActionPath extends AbstractPath
+{
+
+}
