--- src/Services/CodePath/ModelPath.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DevLnk\LaravelCodeBuilder\Services\CodePath;
-
-readonly class ModelPath extends AbstractPath
-{
-}
