--- src/Services/CodePath/RoutePath.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DevLnk\LaravelCodeBuilder\Services\CodePath;
-
-readonly class RoutePath extends AbstractPath
-{
-
-}
